[Patchset 9] trace_processor: surface priority and end state information in sched tables
This allows to see why a process was descheduled and what priority it was
running at.

Bug: 122876520

+++ src/trace_processor/string_table.h
@@ -59,8 +59,8 @@ class StringTable : public Table {
     int Column(sqlite3_context*, int N) override;
 
    private:
-    size_t row_ = 0;
-    size_t num_rows_ = 0;
+    uint32_t row_ = 0;
+    uint32_t num_rows_ = 0;
     const TraceStorage* const storage_;
   };
